requirements.txt
Django==4.1.11
djangorestframework==3.14.0
markdown==3.4.4
django-filter==23.3
mysqlclient==2.2.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.9.1
django-mock-queries==2.2.0
